Move settings realted stuff into feature folder
