+++ app/src/lib/settings/DetailsForm.svelte
