+++ app/src/lib/settings/SupportersBanner.svelte
