+++ app/src/routes/[projectId]/settings/+page.svelte
@@ -1,16 +1,16 @@
 <script lang="ts">
 	import { Project, ProjectService } from '$lib/backend/projects';
 	import BaseBranchSwitch from '$lib/components/BaseBranchSwitch.svelte';
-	import CloudForm from '$lib/components/CloudForm.svelte';
-	import DetailsForm from '$lib/components/DetailsForm.svelte';
-	import KeysForm from '$lib/components/KeysForm.svelte';
-	import PreferencesForm from '$lib/components/PreferencesForm.svelte';
 	import RemoveProjectButton from '$lib/components/RemoveProjectButton.svelte';
 	import SectionCard from '$lib/components/SectionCard.svelte';
 	import Spacer from '$lib/components/Spacer.svelte';
-	import ContentWrapper from '$lib/components/settings/ContentWrapper.svelte';
 	import { featureBaseBranchSwitching } from '$lib/config/uiFeatureFlags';
 	import { showError } from '$lib/notifications/toasts';
+	import CloudForm from '$lib/settings/CloudForm.svelte';
+	import ContentWrapper from '$lib/settings/ContentWrapper.svelte';
+	import DetailsForm from '$lib/settings/DetailsForm.svelte';
+	import KeysForm from '$lib/settings/KeysForm.svelte';
+	import PreferencesForm from '$lib/settings/PreferencesForm.svelte';
 	import { getContext } from '$lib/utils/context';
 	import * as toasts from '$lib/utils/toasts';
 	import { platform } from '@tauri-apps/api/os';
