+++ app/src/routes/settings/telemetry/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-	import AnalyticsSettings from '$lib/components/AnalyticsSettings.svelte';
-	import ContentWrapper from '$lib/components/settings/ContentWrapper.svelte';
+	import AnalyticsSettings from '$lib/settings/AnalyticsSettings.svelte';
+	import ContentWrapper from '$lib/settings/ContentWrapper.svelte';
 </script>
 
 <ContentWrapper title="Telemetry">
